Dynamic sun image, scale an brightness
Sun image changes based on the type of the star, it's size, and the planet distance from the star.
Sky and lighting brightness on a planet changes based on the star size and the distance.
Fixes sun being rendered in front of satellites, but only at airless planets (mainly moons).