Fix sun being rendered in front of satellites
Only fixes scenario for atmosphereless planets, where sun and satellites are present at the same time of day. Might still look weird (as in vanilla) for planets with transparent day sky colors (which is very rare, if even exists in form of mods).

Can't believe this issue is a thing of vanilla sb lmao